modularized the code
