+++ api/user/user.model.js
@@ -1,6 +1,4 @@
 var mongoose = require('mongoose');
-
-
 var userSchema = mongoose.Schema({
     name: {
         type: String,
@@ -12,8 +10,5 @@ var userSchema = mongoose.Schema({
     },
 });
 
-//module.exports = mongoose.model('User', userSchema);
-var User = mongoose.model('User', userSchema);
-module.exports = User;
+module.exports = mongoose.model('User', userSchema);
 
-//NEED SCHEMA FOR SWIM EVENT NAMES!!
