+++ config/mongoose.connection.js
@@ -1,0 +1,4 @@
+var mongoose = require('mongoose');
+module.exports = function(url) {
+    mongoose.connect(url);
+}
