+++ config/routes.express.js
@@ -1,0 +1,14 @@
+var userRouter = require('../api/user/user.router');
+
+module.exports = function(app) {
+    app.use('/users', userRouter);
+
+
+/* This should be last middleware */
+    app.use('*', function(req, res) {
+        res.status(404).json({
+            message: 'Nothing Found.  Youre way off the path.'
+        });
+    });
+
+}
